fix errror nTask option

+++ Samples/UnifiedAPI/ArmoniK.Samples.Client/Program.cs
@@ -23,8 +23,16 @@
 // limitations under the License.
 
 using System;
+using System.CommandLine;
 using System.IO;
 using System.Threading.Tasks;
+
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
+
+using Serilog;
+using Serilog.Events;
+using Serilog.Extensions.Logging;
 
 namespace ArmoniK.Samples.Client
 {
@@ -90,7 +98,7 @@ namespace ArmoniK.Samples.Client
 
                                 numberOfDoubleElement = numberOfBytes == 0
                                                           ? numberOfDoubleElement
-                                                          : numberOfBytes / 8;
+                                                          : (int)(numberOfBytes / 8);
 
                                 var test1 = new LargePayloadTests(configuration_,
                                                                   factory);
